Playing around with some animations, dashboard, upgraded mvvmcross version
