+++ SigfolioWallet.Core/Services/NavigationPath.cs
@@ -1,4 +1,4 @@
 ﻿namespace SigfolioWallet.Core.Services
 {
-    public enum NavigationPath { Home, Transactions, History, Settings }
+    public enum NavigationPath { Home, Transactions, SendReceive, History, Settings }
 }
