+++ SigfolioWallet.Core/ViewModels/SendReceiveViewModel.cs
@@ -1,0 +1,11 @@
+﻿using MvvmCross.ViewModels;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SigfolioWallet.Core.ViewModels
+{
+    public class SendReceiveViewModel : MvxViewModel
+    {
+    }
+}
