+++ SigfolioWallet.UWP.Test/SigfolioWallet.UWP.Test.csproj
@@ -129,7 +129,7 @@
       <Version>6.1.7</Version>
     </PackageReference>
     <PackageReference Include="Moq">
-      <Version>4.9.0</Version>
+      <Version>4.10.0</Version>
     </PackageReference>
     <PackageReference Include="MSTest.TestAdapter">
       <Version>1.3.2</Version>
@@ -138,10 +138,10 @@
       <Version>1.3.2</Version>
     </PackageReference>
     <PackageReference Include="MvvmCross">
-      <Version>6.1.2</Version>
+      <Version>6.2.0</Version>
     </PackageReference>
     <PackageReference Include="MvvmCross.Tests">
-      <Version>6.1.2</Version>
+      <Version>6.2.0</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
