+++ SigfolioWallet.UWP/Utilities/UIUtility.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Windows.UI;
+﻿using Windows.UI;
 using Windows.UI.ViewManagement;
 using Windows.UI.Xaml;
 
-namespace SigfolioWallet
+namespace SigfolioWallet.Utilities
 {
     public static class UIUtility
     {
