+++ SigfolioWallet.UWP/Utilities/UWPUtilities.cs
@@ -1,9 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Media;
 
@@ -20,13 +17,24 @@ namespace SigfolioWallet.Utilities
 
             foreach (var control in controls)
             {
-                if(control.GetValue(FrameworkElement.NameProperty).ToString() == controlName)
+                if (control.GetValue(FrameworkElement.NameProperty).ToString() == controlName)
                 {
                     return control;
                 }
             }
 
             return null;
+        }
+
+        public static T FindParent<T>(DependencyObject dependencyObject) where T : DependencyObject
+        {
+            var parent = VisualTreeHelper.GetParent(dependencyObject);
+
+            if (parent == null)
+                return null;
+
+            var parentT = parent as T;
+            return parentT ?? FindParent<T>(parent);
         }
 
         public static IEnumerable<T> FindVisualChildren<T>(DependencyObject depObj) where T : DependencyObject
@@ -47,6 +55,11 @@ namespace SigfolioWallet.Utilities
                     }
                 }
             }
+        }
+
+        public static T GetValue<T>(this DependencyObject depObj, DependencyProperty dependencyProperty)
+        {
+            return (T)depObj.GetValue(dependencyProperty);
         }
     }
 }
