+++ SigfolioWallet.UWP/Views/HistoryView.xaml.cs
