+++ SigfolioWallet.UWP/Views/HomeView.xaml.cs
@@ -1,26 +1,16 @@
-﻿using MvvmCross.Platforms.Uap.Presenters.Attributes;
+﻿using System;
+using Windows.UI;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Media;
+using MvvmCross.Platforms.Uap.Presenters.Attributes;
 using MvvmCross.Platforms.Uap.Views;
 using MvvmCross.ViewModels;
 using SigfolioWallet.Core.ViewModels;
-using stellar_dotnet_sdk;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace SigfolioWallet
+namespace SigfolioWallet.Views
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
@@ -33,38 +23,26 @@ namespace SigfolioWallet
         {
             this.InitializeComponent();
 
-            LoadAccountDetails();
+
+
+           
+            Loaded += HomeView_Loaded;
         }
 
-        public new HomeViewModel ViewModel => (HomeViewModel)base.ViewModel;
-
-
-        private async void LoadAccountDetails()
+        private void HomeView_Loaded(object sender, RoutedEventArgs e)
         {
-            //var details = await AppShell.server.Accounts.Account(KeyPair.FromAccountId(AppShell.SelectedAccountId));
-
-            ////Currently restricting to XLM only.
-            //var balances = details.Balances.Where(b => b.AssetType == "native").Select(b => new AssetItem()
-            //{
-            //    Asset = b.AssetType == "native" ? "XLM" : b.AssetCode,
-            //    Amount = b.BalanceString
-            //});
-
-            //gvBalances.ItemsSource = balances;
-
+           
         }
 
-        private void gvBalances_ItemClick(object sender, ItemClickEventArgs e)
+        public new HomeViewModel ViewModel => (HomeViewModel) base.ViewModel;
+
+        private void ButtonBase_OnClick(object sender, RoutedEventArgs e)
         {
-            var rowItem = (AssetItem)e.ClickedItem;
+            Random rnd = new Random();
+            var num = rnd.Next(0, 5000);
 
-            this.Frame.Navigate(typeof(TransactionsView), rowItem.Asset);
+            Number.Text = num.ToString();
+            Card1.PrimaryText = num.ToString();
         }
-    }
-
-    class AssetItem
-    {
-        public string Asset { get; set; }
-        public string Amount { get; set; }
     }
 }
