+++ SigfolioWallet.UWP/Views/TransactionsView.xaml
@@ -1,5 +1,5 @@
 ﻿<views:MvxWindowsPage
-    xmlns:views="using:MvvmCross.Platforms.Uap.Views"    x:Class="SigfolioWallet.TransactionsView"
+    xmlns:views="using:MvvmCross.Platforms.Uap.Views"    x:Class="SigfolioWallet.Views.TransactionsView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
